+++ test/test_helper.rb
@@ -3,6 +3,7 @@
 $LOAD_PATH.unshift File.expand_path('../lib', __dir__)
 
 require 'minitest/autorun'
+require 'minitest/mock'
 require 'tmpdir'
 require 'fileutils'
 require 'transpareo/md2pdf'